requirements.txt
certifi==2020.12.5
chardet==4.0.0
DateTime==4.3
idna==2.10
oauthlib==3.1.0
Pillow==8.1.0
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.10.0
urllib3==1.26.3
zope.interface==5.2.0
